move WIN and PYPY tests into pyramid.util
