+++ src/pyramid/config/views.py
@@ -33,7 +33,6 @@ from pyramid.interfaces import (
 from pyramid import renderers
 
 from pyramid.asset import resolve_asset_spec
-from pyramid.compat import WIN
 
 from pyramid.decorator import reify
 
@@ -54,7 +53,12 @@ from pyramid.url import parse_url_overrides
 
 from pyramid.view import AppendSlashNotFoundViewFactory
 
-from pyramid.util import as_sorted_tuple, is_nonstr_iter, TopologicalSorter
+from pyramid.util import (
+    as_sorted_tuple,
+    is_nonstr_iter,
+    TopologicalSorter,
+    WIN,
+)
 
 import pyramid.predicates
 import pyramid.viewderivers
